requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.1.0
dnspython==2.2.1
emoji==1.7.0
idna==3.3
pymongo==4.1.1
PySocks==1.7.1
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
